requirements.txt
django==5.0.4
flask
transbank-sdk
gunicorn
django-axes
django-recaptcha
